moved user-facing files to top level (scripts) (#209)
* moved set_config.py to scripts

* templater.py and atpars_to_jinja2.py to scripts

* python -m pytest in tox.ini

* python -m pylint tests in CodeQC.yaml

* added export PYTHONPATH=$(pwd) pylint tests

* empty __init__.py
