+++ .github/workflows/CodeQC.yaml
@@ -38,4 +38,4 @@ jobs:
 
     - name: Analyzing tests with pylint
       run: |
-        pylint tests
+        export PYTHONPATH=$(pwd) pylint tests
