+++ .github/workflows/pytest-cov.yaml
@@ -26,7 +26,7 @@ jobs:
 
       - name: Run pytest-cov
         run: |
-          pytest --cov-fail-under=70 --cov-report=html --cov=uwtools
+          python -m pytest --cov-fail-under=70 --cov-report=html --cov=uwtools
           rm ./htmlcov/.gitignore
 
       - name: Upload Reports
