+++ scripts/set_config.py
