+++ scripts/templater.py
