+++ tox.ini
@@ -2,6 +2,6 @@
 envlist = py35, py36, py37, py38, py39, py310
 
 [testenv]
-commands = pytest
+commands = python -m pytest
 deps =
     pytest
